+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - RS-439: Snap hooks for new configuration parameters, [PR-628](https://github.com/reductstore/reductstore/pull/628)
-- RS-415: Add check to prevent replication to the same bucket, [PR-629](https://github.com/reductstore/reductstore/pull/629)
+- RS-415: Check to prevent replication to the same bucket, [PR-629](https://github.com/reductstore/reductstore/pull/629)
+- RS-527: POST /:bucket/:entry/q endpoint to query with JSON request, [PR-635](https://github.com/reductstore/reductstore/pull/635)
 
 ### Internal